# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.HTTPError:
    properties:
      error:
        type: string
    type: object
  models.BalanceSheet:
    properties:
      accounts_receivable:
        type: string
      cash:
        type: string
      current_liabilities:
        type: string
      equity:
        type: string
      file_id:
        type: string
      fixed_assets:
        type: string
      future_housing_repairs_payments:
        type: string
      id:
        type: integer
      intangible_assets:
        type: string
      inventories:
        type: string
      investments:
        type: string
      marketable_securities:
        type: string
      non_current_liabilities:
        type: string
      provisions:
        type: string
      statement_id:
        type: string
      total_assets:
        type: string
      total_current_assets:
        type: string
      total_equities:
        type: string
      total_non_current_assets:
        type: string
    type: object
  models.BeneficialOwner:
    properties:
      birth_date:
        type: string
      forename:
        type: string
      id:
        type: integer
      last_modified_at:
        type: string
      latvian_identity_number_masked:
        type: string
      legal_entity_registration_number:
        type: string
      nationality:
        type: string
      registered_on:
        type: string
      residence:
        type: string
      surname:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for managing data based on the provided schema.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your API Title
  version: "1.0"
paths:
  /balance-sheets:
    get:
      description: Retrieve a list of all balance sheet records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BalanceSheet'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all balance sheet entries
      tags:
      - balance-sheets
    post:
      consumes:
      - application/json
      description: Add a new balance sheet record to the database
      parameters:
      - description: Balance Sheet data to create
        in: body
        name: balanceSheet
        required: true
        schema:
          $ref: '#/definitions/models.BalanceSheet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BalanceSheet'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new balance sheet entry
      tags:
      - balance-sheets
  /balance-sheets/{id}:
    delete:
      description: Remove a balance sheet record from the database using its ID
      parameters:
      - description: Balance Sheet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Balance Sheet not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete a balance sheet entry by ID
      tags:
      - balance-sheets
    get:
      description: Retrieve details of a specific balance sheet record using its ID
      parameters:
      - description: Balance Sheet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BalanceSheet'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Balance Sheet not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single balance sheet entry by ID
      tags:
      - balance-sheets
    put:
      consumes:
      - application/json
      description: Modify the details of an existing balance sheet record by ID
      parameters:
      - description: Balance Sheet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Balance Sheet data to update
        in: body
        name: balanceSheet
        required: true
        schema:
          $ref: '#/definitions/models.BalanceSheet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BalanceSheet'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Balance Sheet not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing balance sheet entry
      tags:
      - balance-sheets
  /beneficial-owners:
    get:
      description: Retrieve a list of all beneficial owner records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BeneficialOwner'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all beneficial owner entries
      tags:
      - beneficial-owners
    post:
      consumes:
      - application/json
      description: Add a new beneficial owner record to the database
      parameters:
      - description: Beneficial Owner data to create
        in: body
        name: beneficialOwner
        required: true
        schema:
          $ref: '#/definitions/models.BeneficialOwner'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BeneficialOwner'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new beneficial owner entry
      tags:
      - beneficial-owners
  /beneficial-owners/{id}:
    delete:
      description: Remove a beneficial owner record from the database using its ID
      parameters:
      - description: Beneficial Owner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Beneficial Owner not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete a beneficial owner entry by ID
      tags:
      - beneficial-owners
    get:
      description: Retrieve details of a specific beneficial owner record using its
        ID
      parameters:
      - description: Beneficial Owner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BeneficialOwner'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Beneficial Owner not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single beneficial owner entry by ID
      tags:
      - beneficial-owners
    put:
      consumes:
      - application/json
      description: Modify the details of an existing beneficial owner record by ID
      parameters:
      - description: Beneficial Owner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Beneficial Owner data to update
        in: body
        name: beneficialOwner
        required: true
        schema:
          $ref: '#/definitions/models.BeneficialOwner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BeneficialOwner'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Beneficial Owner not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing beneficial owner entry
      tags:
      - beneficial-owners
...
schemes:
- http
- https
swagger: "2.0"